fix bug where slider wouldn't update correctly

+++ dashboard/router/router.py
@@ -30,7 +30,7 @@ def get_last_irrigation_data():
 
 @router.route('/irrigation/slider', methods=['POST'])
 def set_irrigation_value():
-    value = request.args.get('value', default=None, type=int)
+    value = request.args.get('value', default=None, type=float)
     dc.set_new_optimal_value(value)
     return Response(status=200)
